+++ gendossier.sh
@@ -76,12 +76,15 @@ write_deps() {
       -i "${ROOT}/bazel-genfiles/src/js/nav.soy.js" \
       -i "${ROOT}/bazel-genfiles/src/js/types.soy.js" \
       -i "${ROOT}/src/js/app.js" \
+      -i "${ROOT}/src/js/heap.js" \
       -i "${ROOT}/src/js/main.js" \
       -i "${ROOT}/src/js/nav.js" \
       -i "${ROOT}/src/js/page.js" \
       -i "${ROOT}/src/js/search.js" \
       -i "${ROOT}/src/js/serviceworker/main.js" \
+      -i "${ROOT}/test/js/heap_test.js" \
       -i "${ROOT}/test/js/nav_test.js" \
+      -i "${ROOT}/test/js/search_test.js" \
       -i "${ROOT}/third_party/js/soy/soyutils_usegoog.js" \
       -o "${ROOT}/test/js/deps.js"
 }